TuringGLM

Documentation for TuringGLM.

TuringGLM.center_predictors

TuringGLM.convert_str_to_indices

TuringGLM.data_fixed_effects

TuringGLM.data_random_effects

TuringGLM.data_response

TuringGLM.get_idx

TuringGLM.get_var

TuringGLM.has_ranef

TuringGLM.has_zerocorr

TuringGLM.intercept_per_ranef

TuringGLM.n_ranef

TuringGLM.ranef

TuringGLM.slope_per_ranef

TuringGLM.standardize_predictors

TuringGLM.tuple_length

TuringGLM.center_predictors — Method

center_predictors(X::AbstractMatrix)

Centers the columns of a matrix X of predictors to mean 0.

Returns a tuple with:

μ_X: 1xK Matrix of Float64s of the means of the K columns in the original X

matrix.

X_centered: A Matrix of Float64s with the same dimensions as the original matrix

X with the columns centered on mean μ=0.

Arguments

X::AbstractMatrix: a matrix of predictors where rows are observations and columns are

variables.

TuringGLM.convert_str_to_indices — Method

convert_str_to_indices(v::AbstractVector)

Converts a vector v to a vector of indices, i.e. a vector where all the entries are integers. Returns a tuple with the first element as the converted vector and the second element a Dict specifying which string is which integer.

This function is especially useful for random-effects varying-intercept hierarchical models. Normally v would be a vector of group membership with values such as "group_1", "group_2" etc. For random-effect models with varying-intercepts, Turing needs the group membership values to be passed as Ints.

TuringGLM.data_fixed_effects — Method

data_fixed_effects(formula::FormulaTerm, data)

Constructs the matrix X of fixed-effects (a.k.a. population-level) predictors.

Returns a Matrix of the fixed-effects predictors variables in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.data_random_effects — Method

data_random_effects(formula::FormulaTerm, data)

Constructs the vector(s)/matrix(ces) Z(s) of random-effects (a.k.a. group-level) slope predictors.

Returns a Dict{String, AbstractArray} of Vector/Matrix as values of the random-effects predictors slope variables (keys) in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.data_response — Method

data_response(formula::FormulaTerm, data)

Constructs the response y vector.

Returns a Vector of the response variable in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.get_idx — Method

get_idx(term::Term, data)

Returns a tuple with the first element as the ID vector of Ints that represent group membership for a specific random-effect intercept group t of observations present in data. The second element of the tuple is a Dict specifying which string is which integer in the ID vector.

TuringGLM.get_var — Method

get_var(term::Term, data)

Returns the corresponding vector of column in data for the a specific random-effect slope term of observations present in data.

TuringGLM.has_ranef — Method

has_ranef(formula::FormulaTerm)

Returns true if any of the terms in formula is a FunctionTerm or false otherwise.

TuringGLM.has_zerocorr — Method

has_zerocorr(formula::FormulaTerm)

Returns true if any of the terms in formula is a ZeroCorr or false otherwise.

TuringGLM.intercept_per_ranef — Method

intercept_per_ranef(terms::Tuple{RandomEffectsTerm})

Returns a vector of Strings where the entries are the grouping variables that have a group-level intercept.

TuringGLM.n_ranef — Method

n_ranef(formula::FormulaTerm)

Returns the number of RandomEffectsTerms in formula.

TuringGLM.ranef — Method

ranef(formula::FormulaTerm)

Returns a tuple of the FunctionTerms parsed as RandomEffectsTerms in formula. If there are no FunctionTerms in formula returns nothing.

TuringGLM.slope_per_ranef — Method

slope_per_ranef(terms::Tuple{RandomEffectsTerm})

Returns a SlopePerRanEf object where the entries are the grouping variables that have a group-level slope.

TuringGLM.standardize_predictors — Method

standardize_predictors(X::AbstractMatrix)

Standardizes the columns of a matrix X of predictors to mean 0 and standard deviation 1.

Returns a tuple with:

μ_X: 1xK Matrix of Float64s of the means of the K columns in the original X

matrix.

σ_X: 1xK Matrix of Float64s of the standard deviations of the K columns in the

original X matrix.

X_std: A Matrix of Float64s with the same dimensions as the original matrix

X with the columns centered on mean μ=0 and standard deviation σ=1.

Arguments

X::AbstractMatrix: a matrix of predictors where rows are observations and columns are

variables.

TuringGLM.tuple_length — Method

tuple_length(::NTuple{N, Any}) where {N} = Int(N)

This is a hack to get the length of any tuple.