TuringGLM

Documentation for TuringGLM.

TuringGLM.And1

TuringGLM.AssociativeRule

TuringGLM.CategoricalTerm

TuringGLM.ConstantTerm

TuringGLM.ContinuousTerm

TuringGLM.ContrastsMatrix

TuringGLM.Distributive

TuringGLM.DummyCoding

TuringGLM.FormulaRewrite

TuringGLM.FormulaTerm

TuringGLM.FullDummyCoding

TuringGLM.InteractionTerm

TuringGLM.InterceptTerm

TuringGLM.MatrixTerm

TuringGLM.Schema

TuringGLM.Star

TuringGLM.Term

Base.:==

Base.convert

TuringGLM.apply_schema

TuringGLM.apply_schema

TuringGLM.center_predictors

TuringGLM.checkcol

TuringGLM.checknamesexist

TuringGLM.coefnames

TuringGLM.collect_matrix_terms

TuringGLM.concrete_term

TuringGLM.data_fixed_effects

TuringGLM.data_random_effects

TuringGLM.data_response

TuringGLM.fuzzymatch

TuringGLM.has_schema

TuringGLM.is_matrix_term

TuringGLM.schema

TuringGLM.standardize_predictors

TuringGLM.term

TuringGLM.termsyms

TuringGLM.termvars

TuringGLM.@formula

TuringGLM.And1 — Type

And1 <: FormulaRewrite

Remove numbers from interaction terms, so 1&x becomes &(x) (which is later cleaned up by EmptyAnd).

TuringGLM.AssociativeRule — Type

struct AssociativeRule <: FormulaRewrite end

Apply associative rule: if in an expression headed by an associative operator (+,&,*) and the sub-expression child_idx is headed by the same operator, splice that child's children into it's location.

TuringGLM.CategoricalTerm — Type

CategoricalTerm{C,T,N} <: AbstractTerm

Represents a categorical term, with a name and TuringGLM.ContrastsMatrix

Fields

sym::Symbol: The name of the variable

contrasts::ContrastsMatrix: A contrasts matrix that captures the unique values this variable takes on and how they are mapped onto numerical predictors.

TuringGLM.ConstantTerm — Type

ConstantTerm{T<:Number} <: AbstractTerm

Represents a literal number in a formula. By default will be converted to [TuringGLM.InterceptTerm] by TuringGLM.apply_schema.

Fields

n::T: The number represented by this term.

TuringGLM.ContinuousTerm — Type

ContinuousTerm <: AbstractTerm

Represents a continuous variable, with a name and summary statistics.

Fields

sym::Symbol: The name of the variable

mean::T: Mean

var::T: Variance

min::T: Minimum value

max::T: Maximum value

TuringGLM.ContrastsMatrix — Method

ContrastsMatrix(contrasts::AbstractContrasts, levels::AbstractVector)
ContrastsMatrix(contrasts_matrix::ContrastsMatrix, levels::AbstractVector)

An instantiation of a contrast coding system for particular levels

This type is used internally for generating model matrices based on categorical data, and most users will not need to deal with it directly. Conceptually, a ContrastsMatrix object stands for an instantiation of a contrast coding system for a particular set of categorical data levels.

If levels are specified in the AbstractContrasts, those will be used, and likewise for the base level (which defaults to the first level).

Constructors

ContrastsMatrix(contrasts::AbstractContrasts, levels::AbstractVector)
ContrastsMatrix(contrasts_matrix::ContrastsMatrix, levels::AbstractVector)

Arguments

contrasts::AbstractContrasts: The contrast coding system to use. We only use DummyCoding()

levels::AbstractVector: The levels to generate contrasts for.

contrasts_matrix::ContrastsMatrix: Constructing a ContrastsMatrix from another will check that the levels match. This is used, for example, in constructing a model matrix from a ModelFrame using different data.

TuringGLM.Distributive — Type

struct Distributive <: FormulaRewrite end

Distributive propery: &(a..., +(b...), c...) to +(&(a..., b_i, c...)_i...). Replace outer call (:&) with inner call (:+), whose arguments are copies of the outer call, one for each argument of the inner call. For the ith new child, the original inner call is replaced with the ith argument of the inner call.

TuringGLM.DummyCoding — Type

DummyCoding([base[, levels]])
DummyCoding(; base=nothing, levels=nothing)

Dummy coding generates one indicator column (1 or 0) for each non-base level.

If levels are omitted or nothing, they are determined from the data by calling the levels function on the data when constructing ContrastsMatrix. If base is omitted or nothing, the first level is used as the base.

Columns have non-zero mean and are collinear with an intercept column (and lower-order columns for interactions) but are orthogonal to each other. In a regression model, dummy coding leads to an intercept that is the mean of the dependent variable for base level.

Also known as "treatment coding" or "one-hot encoding".

TuringGLM.FormulaRewrite — Type

abstract type FormulaRewrite end

Formula parsing is expressed as a bunch of expression re-writes, each of which is a subtype of FormulaRewrite. There are two methods that dispatch on these types: applies(ex, child_idx, rule::Type{<:FormulaRewrite}) checks whether the re-write rule needs to be applied at argument child_idx of expression ex, and rewrite!(ex, child_idx, rule::Type{<:FormulaRewrite}) re-writes ex according to rule at position child_idx, and returns the next child_idx that needs to be checked.

TuringGLM.FormulaTerm — Type

FormulaTerm{L,R} <: AbstractTerm

Represents an entire formula, with a left- and right-hand side. These can be of any type (captured by the type parameters).

Fields

lhs::L: The left-hand side (e.g., response)

rhs::R: The right-hand side (e.g., predictors)

TuringGLM.FullDummyCoding — Type

FullDummyCoding()

Full-rank dummy coding generates one indicator (1 or 0) column for each level, including the base level. This is sometimes known as one-hot encoding.

Needed internally for some situations where a categorical variable with $k$ levels needs to be converted into $k$ model matrix columns instead of the standard $k-1$.

TuringGLM.InteractionTerm — Type

InteractionTerm{Ts} <: AbstractTerm

Represents an interaction between two or more individual terms. Generated by combining multiple AbstractTerms with & (which is what calls to & in a @formula lower to)

Fields

terms::Ts: the terms that participate in the interaction.

TuringGLM.InterceptTerm — Type

InterceptTerm{HasIntercept} <: AbstractTerm

Represents the presence or (explicit) absence of an "intercept" term in a regression model. These terms are generated from TuringGLM.ConstantTerms in a formula by TuringGLM.apply_schema(::ConstantTerm, schema). A 1 yields TuringGLM.InterceptTerm{true}, and 0 or -1 yield TuringGLM.InterceptTerm{false} (which explicitly omits an intercept for models which implicitly includes one via the TuringGLM.implicit_intercept trait).

TuringGLM.MatrixTerm — Type

MatrixTerm{Ts} <: AbstractTerm

A collection of terms that should be combined to produce a single numeric matrix.

A matrix term is created by TuringGLM.data_fixed_effects from a tuple of terms using TuringGLM.collect_matrix_terms, which pulls out all the terms that are matrix terms as determined by the trait function TuringGLM.is_matrix_term, which is true by default for all AbstractTerms.

TuringGLM.Schema — Type

Schema

Struct that wraps a Dict mapping Terms to their concrete forms. This exists mainly for dispatch purposes and to support possibly more sophisticated behavior in the future.

A Schema behaves for all intents and purposes like an immutable Dict, and delegates the constructor as well as getindex, get, merge!, merge, keys, and haskey to the wrapped Dict.

TuringGLM.Star — Type

struct Star <: FormulaRewrite end

Expand a*b to a + b + a&b (*(a,b) to +(a,b,&(a,b))). Applies recursively to multiple * arguments, so needs a clean-up pass (from distributive/associative).

TuringGLM.Term — Type

Term <: AbstractTerm

A placeholder for a variable in a formula where the type (and necessary data invariants) is not yet known. This will be converted to a TuringGLM.ContinuousTerm or TuringGLM.CategoricalTerm by TuringGLM.apply_schema.

Fields

sym::Symbol: The name of the data column this term refers to.

Base.:== — Method

Base.:(==)

Only check equality of matrix, termnames, and levels, and that the type is the same for the contrasts (values are irrelevant). This ensures that the two will behave identically in creating modelmatrix columns.

Base.convert — Method

Promote contrasts matrix to full rank version

TuringGLM.apply_schema — Method

apply_schema(t, schema::Schema)

Return a new term that is the result of applying schema to term t.

When t is a ContinuousTerm or CategoricalTerm already, the term will be returned unchanged unless a matching term is found in the schema. This allows selective re-setting of a schema to change the contrast coding or levels of a categorical term, or to change a continuous term to categorical or vice versa.

TuringGLM.apply_schema — Method

apply_schema(t::AbstractTerm, schema::TuringGLM.FullRank)

Apply a schema, under the assumption that when a less-than-full rank model matrix would be produced, categorical terms should be "promoted" to full rank (where a categorical variable with $k$ levels would produce $k$ columns, instead of $k-1$ in the standard contrast coding schemes).

TuringGLM.center_predictors — Method

center_predictors(X::AbstractMatrix)

Centers the columns of a matrix X of predictors to mean 0.

Returns a tuple with:

μ_X: 1xK Matrix of Float64s of the means of the K columns in the original X

matrix.

X_centered: A Matrix of Float64s with the same dimensions as the original matrix

X with the columns centered on mean μ=0.

Arguments

X::AbstractMatrix: a matrix of predictors where rows are observations and columns are

variables.

TuringGLM.checkcol — Method

Return a nice-ish error message if the Symbol name isn't a column name in table, otherwise a zero-length string.

TuringGLM.checknamesexist — Method

Check that each name in the given model f exists in the data source t and return a message if not. Return a zero string otherwise. t is something that implements the Tables interface.

TuringGLM.coefnames — Method

coefnames(term::AbstractTerm)

Return the name(s) of column(s) generated by a term. Return value is either a String or an iterable of Strings.

TuringGLM.collect_matrix_terms — Method

collect_matrix_terms(ts::TupleTerm)
collect_matrix_terms(t::AbstractTerm) = collect_matrix_term((t, ))

Depending on whether the component terms are matrix terms (meaning they have is_matrix_term(T) == true), collect_matrix_terms will return

A single MatrixTerm (if all components are matrix terms)

A tuple of the components (if none of them are matrix terms)

A tuple of terms, with all matrix terms collected into a single MatrixTerm in the first element of the tuple, and the remaining non-matrix terms passed through unchanged.

By default all terms are matrix terms (that is, is_matrix_term(::Type{<:AbstractTerm}) = true), the first case is by far the most common. The others are provided only for convenience when dealing with specialized terms that can't be concatenated into a single model matrix, like random-effects terms.

TuringGLM.concrete_term — Method

concrete_term(t::Term, data[, hint])

Create concrete term from the placeholder t based on a data source and optional hint. If data is a table, the getproperty is used to extract the appropriate column.

The hint can be a Dict{Symbol} of hints, or a specific hint, a concrete term type (ContinuousTerm or CategoricalTerm), or an instance of some <:AbstractContrasts, in which case a CategoricalTerm will be created using those contrasts.

If no hint is provided (or hint==nothing), the eltype of the data is used: Numbers are assumed to be continuous, and all others are assumed to be categorical.

TuringGLM.data_fixed_effects — Method

data_fixed_effects(formula::FormulaTerm, data)

Constructs the matrix X of fixed-effects (a.k.a. population-level) predictors.

Returns a Matrix of the fixed-effects predictors variables in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.data_random_effects — Method

data_random_effects(formula::FormulaTerm, data)

Constructs the matrix(ces) Z(s) of random-effects (a.k.a. group-level) predictors.

Returns a Tuple of Matrix of the random-effects predictors variables in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.data_response — Method

data_response(formula::FormulaTerm, data)

Constructs the response y vector.

Returns a Vector of the response variable in the formula and present inside data.

Arguments

formula: a FormulaTerm created by @formula macro.

data: a data object that satisfies the

Tables.jl interface such as a DataFrame.

TuringGLM.fuzzymatch — Method

This is borrowed from DataFrames.jl. Return between 0 and 2 names from colnames closest to name. colnames : some iterable collection of symbols

TuringGLM.has_schema — Method

has_schema(t::T) where {T<:AbstractTerm}

Return true if t has a schema, meaning that apply_schema would be a no-op.

TuringGLM.is_matrix_term — Method

is_matrix_term(::Type{<:AbstractTerm})

Does this type of term get concatenated with other matrix terms into a single model matrix? This controls the behavior of the TuringGLM.collect_matrix_terms, which collects all of its arguments for which is_matrix_term returns true into a TuringGLM.MatrixTerm, and returns the rest unchanged.

Since all "normal" terms which describe one or more model matrix columns are matrix terms, this defaults to true for any AbstractTerm. An example of a non-matrix term is a random-effect term TuringGLM.RandomEffectTerm.

TuringGLM.schema — Function

schema([terms::AbstractVector{<:AbstractTerm}, ]data, hints::Dict{Symbol})
schema(term::AbstractTerm, data, hints::Dict{Symbol})

Compute all the invariants necessary to fit a model with terms. A schema is a dict that maps Terms to their concrete instantiations (either CategoricalTerms or ContinuousTerms. "Hints" may optionally be supplied in the form of a Dict mapping term names (as Symbols) to term or contrast types. If a hint is not provided for a variable, the appropriate term type will be guessed based on the data type from the data column: any numeric data is assumed to be continuous, and any non-numeric data is assumed to be categorical.

Return a TuringGLM.Schema, which is a wrapper around a Dict mapping Terms to their concrete instantiations (ContinuousTerm or CategoricalTerm).

TuringGLM.standardize_predictors — Method

standardize_predictors(X::AbstractMatrix)

Standardizes the columns of a matrix X of predictors to mean 0 and standard deviation 1.

Returns a tuple with:

μ_X: 1xK Matrix of Float64s of the means of the K columns in the original X

matrix.

σ_X: 1xK Matrix of Float64s of the standard deviations of the K columns in the

original X matrix.

X_std: A Matrix of Float64s with the same dimensions as the original matrix

X with the columns centered on mean μ=0 and standard deviation σ=1.

Arguments

X::AbstractMatrix: a matrix of predictors where rows are observations and columns are

variables.

TuringGLM.term — Method

term(x)

Wrap argument in an appropriate AbstractTerm type: Symbols and AbstractStrings become Terms, and Numbers become ConstantTerms. Any AbstractTerms are unchanged. AbstractStrings are converted to symbols before wrapping.

TuringGLM.termsyms — Method

termsyms(t::Terms.Term)

Extract the set of symbols referenced in this term. This is needed in order to determine when a categorical term should have standard (reduced rank) or full rank contrasts, based on the context it occurs in and the other terms that have already been encountered.

TuringGLM.termvars — Method

termvars(t::AbstractTerm)

The data variables that this term refers to.

TuringGLM.@formula — Macro

@formula(ex)

Capture and parse a formula expression as a Formula struct.

A formula is an abstract specification of a dependence between left-hand and right-hand side variables as in, e.g., a regression model. Each side specifies at a high level how tabular data is to be converted to a numerical matrix suitable for modeling. This specification looks something like Julia code, is represented as a Julia Expr, but uses special syntax. The @formula macro takes an expression like y ~ a*b, transforms it according to the formula syntax rules into a lowered form (like y ~ a + b + a&b), and constructs a Formula struct which captures the original expression, the lowered expression, and the left- and right-hand-side.

Operators that have special interpretations in this syntax are

~ is the formula separator, where it is a binary operator (the first argument is the left-hand side, and the second is the right-hand side.

+ concatenates variables as columns when generating a model matrix.

& representes an interaction between two or more variables, which corresponds to a row-wise kronecker product of the individual terms (or element-wise product if all terms involved are continuous/scalar).

* expands to all main effects and interactions: a*b is equivalent to a+b+a&b, a*b*c to a+b+c+a&b+a&c+b&c+a&b&c, etc.

1, 0, and -1 indicate the presence (for 1) or absence (for 0 and -1) of an intercept column.

The rules that are applied are

The associative rule (un-nests nested calls to +, &, and *).

The distributive rule (interactions & distribute over concatenation +).

The * rule expands a*b to a+b+a&b (recursively).

Subtraction is converted to addition and negation, so x-1 becomes x + -1 (applies only to subtraction of literal 1).

Single-argument & calls are stripped, so &(x) becomes the main effect x.